[ALPS07107800] thermal: fix thermal coverity issue
fix thermal coverity issue

MTK-Commit-Id: 4db8ec11f4a0827f7eddb0fb6d6164a8455406e3

CR-Id: ALPS07107800
Feature: Thermal Management

+++ drivers/misc/mediatek/thermal/common/thermal_zones/mtk_ts_dctm.c
@@ -206,8 +206,8 @@ static int tskinTransient(int tpcb)
 		rhs[i] = div_s64(CapMatrix[i] * tn_1[i], SCALE_UNIT_CAP);
 		tn[i] = 0;
 	}
-	rhs[TSKINNODE] += tamb_coef * tamb;
-	rhs[TPCBNODE]  += tpcb_coef * tpcb;
+	rhs[TSKINNODE] += (long long)tamb_coef * tamb;
+	rhs[TPCBNODE]  += (long long)tpcb_coef * tpcb;
 
 	for (i = 0; i < ACNZ; i++)
 		tn[AcMatrixNzRow[i]] += AcMatrixNz[i] *
